refactor: refactor ending logic
